Refactored goals' structure.
