+++ app/src/main/java/pl/artsit/flexgoals/model/goal/ChoosenQuantitative.java
@@ -1,15 +1,14 @@
-package pl.artsit.flexgoals.model;
+package pl.artsit.flexgoals.model.goal;
 
 import java.util.Date;
 
-public class ChoosenPath {
+public class ChoosenQuantitative {
+    private Integer id;
+    private Integer idGoal;
+    private Integer idUser;
+    private Date date;
 
-   private Integer id;
-   private Integer idGoal;
-   private Integer idUser;
-   private Date date;
-
-    public ChoosenPath(Integer id, Integer idGoal, Integer idUser, Date date) {
+    public ChoosenQuantitative(Integer id, Integer idGoal, Integer idUser, Date date) {
         this.id = id;
         this.idGoal = idGoal;
         this.idUser = idUser;
@@ -50,7 +49,7 @@ public class ChoosenPath {
 
     @Override
     public String toString() {
-        return "choosenPath{" +
+        return "choosenQuantitative{" +
                 "id=" + id +
                 ", idGoal=" + idGoal +
                 ", idUser=" + idUser +
